requirements.txt
OPi.GPIO==0.3.4
pytz==2018.7
PyYAML==3.13
